requirements.txt
streamlit==1.3.0
openai==0.27.0
requests==2.27.1
transformers==4.15.0
Pillow==9.0.1